upnp: factor out code that build a CFileItem from a PLT_MediaObject
